backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# Telegram Bot
python-telegram-bot==20.7

# OCR & PDF Processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# CSV & Data Processing
pandas==2.2.0
chardet==5.2.0

# HTTP Clients for Federation
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
